+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.16.0] - 2026-04-19
+
+
+### Added
+
+- Add webdav support to cloudsync task/credentials (#16)
+
+### Fixed
+
+- Replace deprecated Client().WriteFile calls with service API
+
+### Testing
+
+- Update file resource mocks for WriteFile service API refactor
+
 ## [0.15.2] - 2026-03-25